Make the following updates to customer_segments.ipynb:
Hardcode the three samples selected indices. Use the ones that had been randomly selected in the previous version.
Set random_state=42 in the DecisionTreeRegressor()
Update the new R^2 score of the Grocery DecisionTreeRegressor()
Plot cluster centers together with data means
Add display of probabilities for classification of the three sample points in question 8
Use the probabilities to justify the answer in question 8